refactor(patterns): share cfc helper modules
