+++ packages/patterns/cfc-group-chat-demo/trusted.tsx
@@ -1,7 +1,6 @@
 import {
   type AddIntegrity,
   AuthoredByCurrentUser,
-  Cfc,
   computed,
   Default,
   equals,
@@ -13,7 +12,6 @@ import {
   Stream,
   UI,
   Writable,
-  WriteAuthorizedBy,
 } from "commonfabric";
 import {
   activeAdminRoleForSubject,
@@ -22,7 +20,8 @@ import {
   adminRegistryEntries,
   type EmptyAdminRegistryValue,
   subjectHasAdminRole,
-} from "../admin.ts";
+} from "../cfc/admin.ts";
+import type { TrustedActionWrite } from "../cfc/trusted-action.ts";
 import {
   type ChatProfile,
   type ChatRoom,
@@ -34,23 +33,6 @@ import {
   type PlainChatMessage,
   type SentChatMessage,
 } from "./logic.ts";
-
-export type TrustedActionWrite<
-  T,
-  Binding,
-  Action extends string,
-  Pattern extends string,
-> = Cfc<
-  WriteAuthorizedBy<T, Binding>,
-  {
-    uiContract: {
-      helper: "UiAction";
-      action: Action;
-      trustedPattern: Pattern;
-      requiredEventIntegrity: [Pattern];
-    };
-  }
->;
 
 export const TRUSTED_GROUP_CHAT_PROFILE_SURFACE =
   "TrustedGroupChatProfileSurface";
