+++ packages/patterns/cfc/admin.ts
